+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/persistence/store/PageStore.java
@@ -58,7 +58,7 @@ public interface PageStore extends Closeable {
     /**
      * Returns number of allocated pages.
      */
-    long pages();
+    int pages();
 
     /**
      * Reads a page.
